Remove elastic_ip allocate action and node IP storage
There was an undocumented allocate action that allowed you to store the IP that was returned on the node in a namespace of the resource. We need to stop storing things on the node like they never go away. This is how you end up with EIPs forever (which aren't free). This requires the person to know the IPs now. That's for them to handle in a wrapper cookbook via attribute or via data bag. That might be more work up front, but it's safer and the right way to do it.

Signed-off-by: Tim Smith <tsmith@chef.io>