Fix the wait logic to actually work
Reload Ohai in compile time so it actually reloads during the loop
Use until instead of a loop because exiting out of the loop results in converge_by never firing
Correct the logic that sees if the IP has been added or removed. It was based on count before which makes no sense. Just check to see if the IP is/isn't in the array of private IPs

Signed-off-by: Tim Smith <tsmith@chef.io>